More refactoring and created wiringdiagrams and petrimodels in modeltools
