+++ REQUIRE
@@ -14,3 +14,6 @@ JSON
 Latexify
 Unitful
 Plots
+MacroTools
+Catlab
+ModelingToolkit
